+++ HBridge_ceedling/src/timer.h
@@ -13,7 +13,7 @@ enum deadTimeMultiplier{
 	periodMultiplier_16 = 16
 };
 
-int clk_int_peiord(void);
+float clk_int_peiord(void);
 int getdtgMultiplier(uint32_t deadTime_ns,int period_clk_base);
 int getdtgBitsVal(uint32_t deadTime_ns , int period_clk_base);
 int setDeadTime(uint32_t deadTime_ns);
